+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.0.3
+* Fix ``ConcurrentModificationException`` due to missing synchronized blocks
+* Don't warmUp already warmed up factories again on coordinator level
+* Correctly register factories to ``GlobalPreStartCoordinator``
+* Document ``warmUp``
+* Updated dependencies
+
 # 1.0.2
 * Fix unlikely modification error "Accept exceeded fixed size of ..." during warm-up
 * Updated dependencies